requirements.txt
requests
pymongo
python-dotenv
dnspython